apps/agent/pyproject.toml
[project]
name = "distributed-agent"
version = "0.1.0"
description = "Host Agent for DistributedCompute GPU rental marketplace"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "DistributedCompute Team" }]
keywords = ["gpu", "docker", "agent", "distributed-computing"]

dependencies = [
    "docker>=7.0.0",
    "psutil>=5.9.0",
    "websockets>=12.0",
    "typer>=0.21.0",
    "nvidia-ml-py>=12.0.0",
    "speedtest-cli>=2.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.9.0",
    "structlog>=23.2.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pyinstaller>=6.3.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
distributed-agent = "distributed_agent.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/distributed_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
